move ConditionListAttribute to manager
